+++ sentry/src/main/java/io/sentry/SpanContext.java
@@ -44,8 +44,9 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
   /** A map or list of tags for this event. Each tag must be less than 200 characters. */
   protected @NotNull Map<String, @NotNull String> tags = new ConcurrentHashMap<>();
 
-  /** Describes the status of the Transaction. */
   protected @Nullable String origin = DEFAULT_ORIGIN;
+
+  protected @NotNull Map<String, Object> data = new ConcurrentHashMap<>();
 
   private @Nullable Map<String, Object> unknown;
 
@@ -230,6 +231,14 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
     return baggage;
   }
 
+  public @NotNull Map<String, Object> getData() {
+    return data;
+  }
+
+  public void setData(final @NotNull String key, final @NotNull Object value) {
+    data.put(key, value);
+  }
+
   @ApiStatus.Internal
   public SpanContext copyForChild(
       final @NotNull String operation,
@@ -275,6 +284,7 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
     public static final String STATUS = "status";
     public static final String TAGS = "tags";
     public static final String ORIGIN = "origin";
+    public static final String DATA = "data";
   }
 
   @Override
@@ -301,6 +311,9 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
     }
     if (!tags.isEmpty()) {
       writer.name(JsonKeys.TAGS).value(logger, tags);
+    }
+    if (!data.isEmpty()) {
+      writer.name(JsonKeys.DATA).value(logger, data);
     }
     if (unknown != null) {
       for (String key : unknown.keySet()) {
@@ -336,6 +349,7 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
       SpanStatus status = null;
       String origin = null;
       Map<String, String> tags = null;
+      Map<String, Object> data = null;
 
       Map<String, Object> unknown = null;
       while (reader.peek() == JsonToken.NAME) {
@@ -366,6 +380,9 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
             tags =
                 CollectionUtils.newConcurrentHashMap(
                     (Map<String, String>) reader.nextObjectOrNull());
+            break;
+          case JsonKeys.DATA:
+            data = (Map<String, Object>) reader.nextObjectOrNull();
             break;
           default:
             if (unknown == null) {
@@ -404,9 +421,15 @@ public class SpanContext implements JsonUnknown, JsonSerializable {
       spanContext.setDescription(description);
       spanContext.setStatus(status);
       spanContext.setOrigin(origin);
+
       if (tags != null) {
         spanContext.tags = tags;
       }
+
+      if (data != null) {
+        spanContext.data = data;
+      }
+
       spanContext.setUnknown(unknown);
       reader.endObject();
       return spanContext;